:umbrella: :ambulance: Fix unit tests
