+++ pkg/cmds/parameters/parameters.go
@@ -1,4 +1,4 @@
-package cmds
+package parameters
 
 import (
 	"bufio"
