+++ pkg/cmds/parameters/parameters_test.go
@@ -1,4 +1,4 @@
-package cmds
+package parameters
 
 import (
 	_ "embed"
